requirements.txt
imutils==0.5.4
numpy==2.2.5
opencv-python==4.11.0.86
RPi.GPIO==0.7.1